Fix multiple compile errors on older gcc versions

+++ modules/juce_box2d/box2d/Dynamics/Contacts/b2Contact.h
@@ -239,7 +239,7 @@ inline void b2Contact::SetEnabled(bool flag)
 	}
 	else
 	{
-		m_flags &= ~e_enabledFlag;
+		m_flags &= ~((unsigned int) e_enabledFlag);
 	}
 }
